% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRCT.estimators.R
\name{genRCT.estimators}
\alias{genRCT.estimators}
\title{Estimate the ATE using specified estimators}
\usage{
genRCT.estimators(
  Y.trial,
  A.trial,
  X.trial,
  Y.rwe,
  A.rwe,
  X.rwe,
  family = "gaussian",
  estimators,
  sieve = TRUE,
  seed = NULL,
  osel1.t = NULL,
  osel0.t = NULL,
  osel1.b = NULL,
  osel0.b = NULL,
  osel.ipsw = NULL,
  flag.boot = FALSE
)
}
\arguments{
\item{Y.trial}{Observed outcome from a trial; vector of size \code{n} (trial sample size).}

\item{A.trial}{Treatment received from a trial; vector of size \code{n}.}

\item{X.trial}{Matrix of \code{p} baseline covariates from a trial; dimension \code{n} by \code{p}.}

\item{Y.rwe}{Observed outcome from OS; if obtained, vector of size \code{m} (OS sample size);
otherwise, set \code{Y.rwe = NULL}.}

\item{A.rwe}{Treatment received from OS; if obtained, vector of size \code{m}; otherwise, set \code{A.rwe = NULL}.}

\item{X.rwe}{Matrix of \code{p} baseline covariates from OSl; dimension \code{m} by \code{p}.}

\item{family}{The type of outcome; \code{'gaussian'} for gaussian regression or \code{'binomial'} for logistic regression.
Default is \code{"gaussian"}.}

\item{estimators}{A vector of one or multiple methods to estimate the ATE. Allowed values are
\code{'Naive'}, \code{'IPSW'}, \code{'AIPSW'}, \code{'CW'}, \code{'ACW-t'}, \code{'ACW-b'}.
The \code{'ACW-b'} is allowed only when both \code{Y.rwe} and \code{A.rwe} are obtained.
Default specifies all 6 methods.}

\item{sieve}{A logical value indicating whether the method of sieves are used for estimating sampling score and outcome models.
Used only if \code{estimators = 'AIPSW'} or \code{'ACW-t'} or \code{'ACW-b'}. Default is \code{TRUE}.}

\item{seed}{An optional integer specifying an initial randomization seed for reproducibility.
Default is \code{NULL}, corresponding to no seed.}

\item{osel1.t}{A vector indicating selected outcome model covariates for the \code{'ACW-t'} estimator in the group trt = 1.
This is only relevant when \code{sieve = TRUE} and \code{inference = TRUE}. Otherwise, set \code{osel1.t = NULL}.}

\item{osel0.t}{A vector indicating selected outcome model covariates for the \code{'ACW-t'} estimator in the group trt = 0.
This is only relevant when \code{sieve = TRUE} and \code{inference = TRUE}. Otherwise, set \code{osel0.t = NULL}.}

\item{osel1.b}{A vector indicating selected outcome model covariates for the \code{'ACW-b'} estimator in the group trt = 1.
This is only relevant when \code{sieve = TRUE} and \code{inference = TRUE}. Otherwise, set \code{osel1.b = NULL}.}

\item{osel0.b}{A vector indicating selected outcome model covariates for the \code{'ACW-b'} estimator in the group trt = 0.
This is only relevant when \code{sieve = TRUE} and \code{inference = TRUE}. Otherwise, set \code{osel0.b = NULL}.}

\item{osel.ipsw}{A vector indicating selected sampling scole model covariates for the \code{'AIPSW'} estimator.
This is only relevant when \code{sieve = TRUE} and \code{inference = TRUE}. Otherwise, set \code{osel.ipsw = NULL}.}

\item{flag.boot}{A logical value indicating whether the function estimates bootstrap sample ATEs. Default is \code{FALSE}.}
}
\value{
Return a list containing:\tabular{ll}{
   \code{ate} \tab\tab  A vector of estimated ATEs using the selected methods. \cr
   \tab \cr
   \code{hypers} \tab\tab A list of vectors indicating the selected covariates. \cr
}
}
\description{
ATE estimated with Naive, IPSW, AIPSW, CW, ACW-t, ACW-b.
}
