% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRCT.R
\name{genRCT}
\alias{genRCT}
\title{Generalizing the average treatment effect (ATE) from the trial using observational studies (OS)}
\usage{
genRCT(
  Y.trial,
  A.trial,
  X.trial,
  Y.rwe,
  A.rwe,
  X.rwe,
  family = "gaussian",
  estimators = c("Naive", "IPSW", "AIPSW", "CW", "ACW-t", "ACW-b"),
  sieve = TRUE,
  inference = TRUE,
  n.boot = 100,
  conf.level = 0.05,
  seed = NULL,
  plot.boot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{Y.trial}{Observed outcome from a trial; vector of size \code{n} (the trial sample size).}

\item{A.trial}{Treatment received from a trial; vector of size \code{n}.}

\item{X.trial}{Matrix of \code{p} baseline covariates from a trial; dimension \code{n} by \code{p}.}

\item{Y.rwe}{Observed outcome from OS; if obtained, vector of size \code{m} (OS sample size);
otherwise, set \code{Y.rwe = NULL}.}

\item{A.rwe}{Treatment received from OS; if obtained, vector of size \code{m}; otherwise, set \code{Y.rwe = NULL}.}

\item{X.rwe}{Matrix of \code{p} baseline covariates from OS; dimension \code{m} by \code{p}.}

\item{family}{The type of outcome; \code{"gaussian"} for continuous outcome or \code{"binomial"} for binary outcome.
Default is \code{"gaussian"}.}

\item{estimators}{A vector of one or multiple methods to estimate the ATE. Allowed values are
\code{'Naive'}, \code{'IPSW'}, \code{'AIPSW'}, \code{'CW'}, \code{'ACW-t'}, \code{'ACW-b'}.
The \code{'ACW-b'} is allowed only when both \code{Y.rwe} and \code{A.rwe} are obtained.
Default specifies all 6 methods.}

\item{sieve}{A logical value indicating whether the method of sieves are used for estimating sampling score and outcome models.
Used only if \code{estimators = 'AIPSW'} or \code{'ACW-t'} or \code{'ACW-b'}. Default is \code{TRUE}.}

\item{inference}{A logical value indicating whether inference for the ATE via bootstrap should be provided.
Default it \code{TRUE}.}

\item{n.boot}{A numeric value indicating the number of bootstrap samples used. This is only relevant
if \code{inference = TRUE}. Default is \code{100}.}

\item{conf.level}{The level of bootstrap confidence interval; Default is \code{0.05}.}

\item{seed}{An optional integer specifying an initial randomization seed for reproducibility.
Default is \code{NULL}, corresponding to no seed.}

\item{plot.boot}{A logical value indicating whether histograms of the bootstrap samples should be produced. Default is \code{TRUE}.}

\item{verbose}{A logical value indicating whether intermediate progress messages should be printed.
Default is \code{TRUE}.}
}
\value{
Return a list containing:\tabular{ll}{
   \code{fit} \tab  A table of estimated ATEs with bootstrap SE and confidence interval. \cr
   \tab \cr
   \code{plot} \tab A set of histograms displaying the distribution of the bootstrapped estimates. The red vertical reference lines
    represent the estimated ATEs from each method. \cr
}
}
\description{
Provides estimation and inference for the average treatment effect (ATE), generalizing from a randomized controlled trial (RCT)
to a target population leveraging observational studies. The method of sieves can be used for the estimation of sampling score and outcome models.
The ATE can be estimated with up to 6 methods among the following: Naive, IPSW, AIPSW, CW, ACW-t, ACW-b.
}
\examples{
fit <- genRCT(Y.trial = Y.trial, A.trial = A.trial, X.trial = X.trial,Y.rwe = Y.rwe, A.rwe = A.rwe, X.rwe = X.rwe,
family = "gaussian", estimators = c("Naive", "IPSW", "AIPSW", "CW", "ACW-t", "ACW-b"), sieve = TRUE,
inference = TRUE, n.boot = 500, conf.level = 0.05, seed = 123, plot.boot = TRUE, verbose = TRUE)

}
