% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genRCT.surv.R
\name{genRCT.surv}
\alias{genRCT.surv}
\title{Generalizing the average treatment effect (ATE) from the trial using observational studies (OS) for survival outcomes}
\usage{
genRCT.surv(
  Y.trial,
  d.trial,
  A.trial,
  X.trial,
  X.rwe,
  tau,
  eta.vec = c(0, 0.001, 0.005),
  n.boot = 100,
  conf.level = 0.05,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Y.trial}{Observed outcome from a trial; vector of size \code{n} (trial sample size).}

\item{d.trial}{The event indicator, normally 1 = event, 0 = censored; vector of size \code{n}.}

\item{A.trial}{Treatment received from a trial; vector of size \code{n}.}

\item{X.trial}{Matrix of \code{p} baseline covariates from a trial; dimension \code{n} by \code{p}.}

\item{X.rwe}{Matrix of \code{p} baseline covariates from OS; dimension \code{m} by \code{p}.}

\item{tau}{| A vector of truncation time for defining restricted mean survival time; e.g., seq(10, 50, by = 10)}

\item{n.boot}{A numeric value indicating the number of bootstrap samples used. This is only relevant
if \code{inference = TRUE}. Default is \code{100}.}

\item{conf.level}{The level of bootstrap confidence interval; Default is \code{0.05}.}

\item{seed}{An optional integer specifying an initial randomization seed for reproducibility.
Default is \code{NULL}, corresponding to no seed.}

\item{verbose}{A logical value indicating whether intermediate progress messages should be printed.
Default is \code{TRUE}.}
}
\value{
Return a list containing:\tabular{ll}{
   \code{rmst} \tab\tab  A list of estimated RMSTs with bootstrap SE and confidence interval. \cr
   \tab \cr
   \code{surv} \tab\tab  A list of estimated treatment-specific survival function. \cr
}
}
\description{
Provides estimation and inference for the average treatment effect (ATE), generalizing from a randomized controlled trial (RCT)
to a target population leveraging observational studies. The method of sieves can be used for the estimation of sampling score and outcome models.
}
\examples{
fit <- genRCT.surv(Y.trial = Y.trial, d.trial = d.trial, A.trial = A.trial, X.trial = X.trial, X.rwe = X.rwe,
tau = tau. n.boot = 100, conf.level = 0.05, seed = 123, verbose = TRUE)
}
